refactor: 상수 또는 리터럴 문자를 객체 중 결합도가 높은 쪽으로 이동(전체적으로 많이 사용하는 경우 공통 클래스를 사용)
